+++ src/interfaces/mod.rs
@@ -5,5 +5,8 @@ pub mod statics;
 #[cfg(feature = "ffi")]
 pub mod ffi;
 
+#[cfg(feature = "python")]
+pub mod pyo3;
+
 #[cfg(feature = "wasm_bindgen")]
 pub mod wasm_bindgen;